stm32mp1: change stack to uint32_t to keep 4 bytes aligned
Some crash may occur if the stack is not properly aligned.
By switching to uint32_t, we ensure this.

Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/164260

+++ plat/st/stm32mp1/stm32mp1_low_power.c
@@ -36,7 +36,7 @@
 static unsigned int gicc_pmr;
 static struct stm32_rtc_calendar sleep_time;
 static bool enter_cstop_done;
-static uint8_t int_stack[STM32MP_INT_STACK_SIZE];
+static uint32_t int_stack[STM32MP_INT_STACK_SIZE];
 
 extern void wfi_svc_int_enable(uintptr_t stack_addr);
